Refactor: Extract Reader components
This commit extracts several components from `ReaderScreen.kt` into their own files:

- `ReaderPage.kt`: Contains the main content rendering logic for a single reader page.
- `ReaderViewBottomBar.kt`: Encapsulates the bottom action bar for the reader.
- `ReaderCustomizationsContent.kt`: Manages the UI for reader customization options (renamed from `ReaderCustomisations.kt`).

The `ReaderScreen.kt` has been updated to use these new components, and the original `ReaderCustomisations.kt` file has been removed.
